+++ app/Http/Controllers/ConsigneeControllers/CDispatchTicketController.php
@@ -3,8 +3,13 @@
 namespace App\Http\Controllers\ConsigneeControllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Response;
 use App\Http\Controllers\Controller;
-
+use App\Company;
+use App\Contract;
+use App\JobOrder;
+use DB;
+use Auth;
 class CDispatchTicketController extends Controller
 {
     /**
@@ -13,8 +18,27 @@ class CDispatchTicketController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        return view('Consignee.DispatchTicket.index');
+    {   
+        $dispatch = DB::table('tbljoborder as joborder')
+        ->join('tblservices as service','joborder.intJOServiceTypeID','service.intServicesID')
+        ->join('tblberth as berth','joborder.intJOBerthID','berth.intBerthID')
+        ->join('tblpier as pier','berth.intBPierID','pier.intPierID')
+        // ->join('tblbarge as barge','joborder.intJOBargeID','barge.intBargeID')
+        ->join('tblgoods as goods','joborder.intJOGoodsID','goods.intGoodsID')
+        // ->join('tblvessel as vessel','joborder.intJOeVesselID','vessel.intVesselID')
+        ->join('tblcompany as company','joborder.intJOCompanyID','company.intCompanyID')
+        ->join('tbljobsched as jobsched','joborder.intJobOrderID','jobsched.intJSJobOrderID')
+        ->join('tbltugboatassign as tugboatassign','jobsched.intJSTugboatAssignID','tugboatassign.intTAssignID')
+        ->join('tbltugboat as tugboat','tugboatassign.intTATugboatID','tugboat.intTugboatID')
+        ->join('tbltugboatmain as main','tugboat.intTTugboatMainID','main.intTugboatMainID')
+        ->join('tbldispatchticket as dispatch','dispatch.intDTTugboatAssignID','tugboatassign.intTAssignID')
+        ->where('company.intCompanyID',Auth::user()->intUCompanyID)
+        ->where('dispatch.boolAApprovedby',1)
+        ->where('jobsched.enumstatus','Finished')
+        ->get(); 
+
+        return view('Consignee.Dispatch.index')
+        ->with('dispatch',$dispatch);
     }
 
     /**
@@ -35,7 +59,14 @@ class CDispatchTicketController extends Controller
      */
     public function store(Request $request)
     {
-        //
+            $Dispatch = DispatchTicket::find($request->dispatch);
+            error_log('aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa');
+            error_log($request->dispatch);
+            $Dispatch->timestamps = false;
+            $Dispatch->boolCApprovedby = 1;
+            // return response()->json(['dispatch'=>$dispatch]);    
+            $Dispatch->save();
+            
     }
 
     /**
@@ -82,4 +113,28 @@ class CDispatchTicketController extends Controller
     {
         //
     }
+    public function info($intDispatchTicketID)
+    {
+        $Dispatch = DB::table('tbljoborder as joborder')
+        ->join('tblservices as service','joborder.intJOServiceTypeID','service.intServicesID')
+        ->join('tblberth as berth','joborder.intJOBerthID','berth.intBerthID')
+        ->join('tblpier as pier','berth.intBPierID','pier.intPierID')
+        // ->join('tblbarge as barge','joborder.intJOBargeID','barge.intBargeID')
+        ->join('tblgoods as goods','joborder.intJOGoodsID','goods.intGoodsID')
+        // ->join('tblvessel as vessel','joborder.intJOeVesselID','vessel.intVesselID')
+        ->join('tblcompany as company','joborder.intJOCompanyID','company.intCompanyID')
+        ->join('tbljobsched as jobsched','joborder.intJobOrderID','jobsched.intJSJobOrderID')
+        ->join('tbltugboatassign as tugboatassign','jobsched.intJSTugboatAssignID','tugboatassign.intTAssignID')
+        ->join('tbltugboat as tugboat','tugboatassign.intTATugboatID','tugboat.intTugboatID')
+        ->join('tbltugboatmain as main','tugboat.intTTugboatMainID','main.intTugboatMainID')
+        ->join('tbldispatchticket as dispatch','dispatch.intDTTugboatAssignID','tugboatassign.intTAssignID')
+        ->where('company.intCompanyID',Auth::user()->intUCompanyID)
+        ->where('jobsched.enumstatus','Finished')
+        ->where('dispatch.intDispatchTicketID',$intDispatchTicketID)
+        ->get(); 
+
+        return response()->json(['dispatch'=>$Dispatch]);    
+
+    }
+
 }